+++ packages/blocks-editor/src/Editor/Commands.ts
@@ -1,0 +1,5 @@
+import {createCommand, LexicalCommand} from 'lexical';
+
+export const INSERT_FILE_COMMAND: LexicalCommand<string> = createCommand(
+  'INSERT_FILE_COMMAND',
+);