+++ packages/web/src/javascripts/Utils/Items/Search/ItemLink.ts
@@ -1,0 +1,7 @@
+import { LinkableItem } from './LinkableItem'
+
+export type ItemLink<ItemType extends LinkableItem = LinkableItem> = {
+  id: string
+  item: ItemType
+  type: 'linked' | 'linked-by'
+}