+++ packages/web/src/javascripts/Utils/Items/Search/LinkableItem.ts
@@ -1,0 +1,3 @@
+import { DecryptedItemInterface, ItemContent } from '@standardnotes/snjs'
+
+export type LinkableItem = DecryptedItemInterface<ItemContent>