+++ packages/blocks-editor/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8](https://github.com/standardnotes/app/compare/@standardnotes/blocks-editor@1.3.7...@standardnotes/blocks-editor@1.3.8) (2022-11-08)
+
+**Note:** Version bump only for package @standardnotes/blocks-editor
+
 ## 1.3.7 (2022-11-07)
 
 **Note:** Version bump only for package @standardnotes/blocks-editor