+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.23.283](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.282...@standardnotes/desktop@3.23.283) (2022-11-08)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.23.282](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.281...@standardnotes/desktop@3.23.282) (2022-11-07)
 
 **Note:** Version bump only for package @standardnotes/desktop